Changed nopandora to require manual intervention so it's more reliable.
Automatic detection of pandora's availability doesn't seem to work for all users. We get lots of false negatives. So instead, we make the default look like the noscript version (mostly) and insert the warning message manually when we *know* that pandora is not running: Remove 'NOPANDORA' string from index.html.{en,de}.